requirements/base.txt
Django==1.10.6
django-webpack-loader==0.4.1
django-crispy-forms==1.6.1
django-crontab==0.7.1
django-filter==1.0.1
django-picklefield==0.3.2
djangorestframework==3.6.2

celery==3.1.24
psycopg2==2.7.1
redis==2.10.5

bokeh==0.12.4

git+ssh://git@bitbucket.org/geniussk/bayesian-bmd-stats.git@0.0.7
